refactor: custom hooks for quiz & question
