+++ src/components/QuestionCard/QuestionCard.tsx
@@ -1,7 +1,7 @@
 import "./QuestionCard.css";
 import { QuestionPropType } from "../../context/QuizContext/quiz.types";
 import { Option } from "./Option";
-import { useQuiz } from "../../hooks";
+import { useQuestionHelper } from "./useQuestionHelper";
 
 export const QuestionCard = ({
   question,
@@ -9,43 +9,13 @@ export const QuestionCard = ({
   loading,
   setLoading,
 }: QuestionPropType): JSX.Element => {
-  const { currentQuizAnswer, quizDispatch } = useQuiz();
-
-  const isAlreadyAnswered = (questionId: string): boolean =>
-    currentQuizAnswer.answerList.some(
-      (question) => question.questionId === questionId
-    );
-
-  const selectOptionOnClick = (id: string, isRight: boolean): void => {
-    !loading &&
-      !isAlreadyAnswered(question._id) &&
-      quizDispatch({
-        type: "SET_SCORE",
-        payload: {
-          score: isRight
-            ? currentQuizAnswer.score + question.positiveMarks
-            : currentQuizAnswer.score - question.negativeMarks,
-        },
-      });
-
-    !loading &&
-      !isAlreadyAnswered(question._id) &&
-      quizDispatch({
-        type: "SAVE_ANSWER",
-        payload: {
-          answer: {
-            questionId: question._id,
-            selectedOptionId: id,
-            isCorrect: isRight,
-          },
-        },
-      });
-    setLoading(true);
-    setTimeout(() => {
-      loadNextQuestion();
-      setLoading(false);
-    }, 1500);
-  };
+  
+  const { selectOptionOnClick } = useQuestionHelper({
+    question,
+    loadNextQuestion,
+    loading,
+    setLoading,
+  });
 
   return (
     <div>
