+++ src/pages/UndefinedRoute.tsx
@@ -1,8 +1,14 @@
+import { Link } from "react-router-dom";
+
 export const UndefinedRoute = (): JSX.Element => {
   return (
-    <div>
+    <div
+      style={{ display: "flex", flexDirection: "column", alignItems: "center" }}
+    >
       <h1>404</h1>
-      <h3>Wrong URL, go to home page.</h3>
+      <h3>
+        Wrong URL, go to <Link to="/">Home</Link> page.
+      </h3>
     </div>
   );
 };
